add capability to download all weights at build time
Fixes issue #1. With slow Internet connections, having to download the
weights of ResNet50 and keras-resnet each time the container is run,
becomes a pain. To avoid this, we modify:

- notebooks/COCO_evaluation.ipynb, so that it checks if a model file
exists before downloading it
- README.md, explaining how to download weights just once at build time

and add:

 - Dockerfile.download.weights
 - Container-Root/download_weights.py

which download the weights once at build time. The default behavior
(Dockerfile) remains to download weights at run time, because it's
instructive for beginners to see that many Deep Learning
frameworks/projects download weights from GitHub.